players now require continuous key presses to move
I implemented this differently than Stanley.  When checking for called
actions, if none are called and the current state.action is not Jump,
then the player is stopped by executing Stop.

I also fixed the AI after this, since the continuous key pressing wasn't
set up right, but now it works, pressing whatever keys it pressed for
its cycle length.